versão 1 de UI e items

+++ src/main/java/me/brzeph/core/domain/entity/item/InventoryItem.java
@@ -5,13 +5,13 @@ import me.brzeph.core.domain.gui.core.others.Color;
 // Novo arquivo (ou no mesmo pacote dos itens)
 public final class InventoryItem {
     private final ItemDefinition definition;
-    private final int quantity;
+    private final float quantity;
 
-    public InventoryItem(ItemDefinition definition, int quantity){
+    public InventoryItem(ItemDefinition definition, float quantity){
         this.definition = definition;
         this.quantity = Math.max(1, quantity);
     }
 
     public ItemDefinition definition(){ return definition; }
-    public int quantity(){ return quantity; }
+    public float quantity(){ return quantity; }
 }
